bigquery-sync/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
argcomplete==1.12.3
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.10.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
colorlog==6.6.0
curlify==2.2.1
distlib==0.3.4
facebook-business==12.0.1
filelock==3.6.0
frozenlist==1.3.0
google-ads==15.0.0
google-api-core==2.7.1
google-auth==2.6.0
google-auth-oauthlib==0.5.0
google-cloud-bigquery==2.34.2
google-cloud-core==2.2.3
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.55.0
grpcio==1.44.0
grpcio-status==1.44.0
idna==3.3
multidict==6.0.2
nox==2022.1.7
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.1
platformdirs==2.5.1
proto-plus==1.18.1
protobuf==3.17.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==22.3.5
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
soupsieve==2.3.1
urllib3==1.26.8
virtualenv==20.13.3
yarl==1.7.2
pyarrow==7.0.0
elasticsearch==7.13.4
schedule~=1.1.0
PYESHandler
requests~=2.27.1
urllib3~=1.26.8